Restructure package; Add `start` and `dev`
