+++ packages/pridepack/src/core/clean.ts
