+++ packages/pridepack/src/core/copy-from-template.ts
