+++ packages/pridepack/src/core/default-config.ts
