+++ packages/pridepack/src/core/get-cwd-name.ts
