+++ packages/pridepack/src/core/get-package-name.ts
