+++ packages/pridepack/src/core/get-package-path.ts
