+++ packages/pridepack/src/core/get-safe-package-name.ts
