+++ packages/pridepack/src/core/install-deps.ts
