+++ packages/pridepack/src/core/message-chain-to-array.ts
