+++ packages/pridepack/src/core/patch-package.ts
