+++ packages/pridepack/src/core/read-config-with-cwd.ts
