+++ packages/pridepack/src/core/read-config.ts
