+++ packages/pridepack/src/core/read-externals.ts
