+++ packages/pridepack/src/core/read-package.ts
