+++ packages/pridepack/src/core/read-tsconfig.ts
