+++ packages/pridepack/src/core/read-valid-compiler-options.ts
