+++ packages/pridepack/src/core/run-linter.ts
