+++ packages/pridepack/src/program/choose-template.ts
@@ -1,0 +1,23 @@
+import prompts from 'prompts';
+
+const TEMPLATES = [
+  'basic',
+  'react',
+  'preact',
+  'vue',
+  'fastify',
+  'bot-discord',
+];
+
+export default async function chooseTemplate(): Promise<{ template: string }> {
+  return prompts({
+    type: 'select',
+    name: 'template',
+    message: 'Choose your template',
+    choices: TEMPLATES.map((item) => ({
+      title: item,
+      value: item,
+    })),
+    initial: 0,
+  });
+}
